refactor(aia): move indirect CSR accessors to individual modules
