+++ library/riscv-aia/src/csrind.rs
@@ -1,0 +1,6 @@
+//! Indirect CSRs for RISC-V AIA.
+
+pub mod eidelivery;
+pub mod eie;
+pub mod eip;
+pub mod eithreshold;
