+++ library/riscv-aia/src/peripheral/imsic.rs
@@ -18,169 +18,6 @@ pub struct Imsic {
     pub seteipnum_be: WO<u32>,
 }
 
-impl Imsic {
-    /// Returns the size of the IMSIC register block.
-    pub const fn size() -> usize {
-        0x1000 // 4-KiB page
-    }
-}
-
-/// External interrupt delivery enable register (`eidelivery`).
-///
-/// This register controls whether interrupts from this interrupt file are
-/// delivered from the IMSIC to the attached hart.
-///
-/// *NOTE:* Guest interrupt files do not support value 0x40000000 for `eidelivery`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct Eidelivery(u32);
-
-impl Eidelivery {
-    /// Interrupt delivery is disabled.
-    pub const DISABLED: Self = Self(0);
-
-    /// Interrupt delivery from the interrupt file is enabled.
-    pub const ENABLED: Self = Self(1);
-
-    /// Interrupt delivery from a PLIC or APLIC is enabled (optional).
-    pub const PLIC_APLIC_ENABLED: Self = Self(0x40000000);
-
-    /// Creates a new `Eidelivery` from a raw value.
-    pub const fn from_raw(value: u32) -> Self {
-        Self(value)
-    }
-
-    /// Returns the raw value of this register.
-    pub const fn raw(self) -> u32 {
-        self.0
-    }
-
-    /// Returns whether interrupt delivery is disabled.
-    pub const fn is_disabled(self) -> bool {
-        self.0 == Self::DISABLED.0
-    }
-
-    /// Returns whether interrupt delivery from the interrupt file is enabled.
-    pub const fn is_enabled(self) -> bool {
-        self.0 == Self::ENABLED.0
-    }
-
-    /// Returns whether interrupt delivery from a PLIC or APLIC is enabled.
-    pub const fn is_plic_aplic_enabled(self) -> bool {
-        self.0 == Self::PLIC_APLIC_ENABLED.0
-    }
-}
-
-/// External interrupt threshold register (`eithreshold`).
-///
-/// This register specifies the minimum priority level for an interrupt
-/// to be delivered. Interrupts with priority below this threshold are masked.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct Eithreshold(u32);
-
-impl Eithreshold {
-    /// Creates a new `Eithreshold` from a raw value.
-    pub const fn from_raw(value: u32) -> Self {
-        Self(value)
-    }
-
-    /// Returns the raw value of this register.
-    pub const fn raw(self) -> u32 {
-        self.0
-    }
-
-    /// Returns the threshold value.
-    pub const fn threshold(self) -> u32 {
-        self.0
-    }
-
-    /// Returns the priority threshold.
-    pub const fn priority(self) -> u8 {
-        (self.0 >> 24) as u8
-    }
-}
-
-/// External interrupt-pending bits register (`eip[n]`).
-///
-/// Each bit represents whether the corresponding interrupt identity is pending.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct Eip(u32);
-
-impl Eip {
-    /// Creates a new `Eip` from a raw value.
-    pub const fn from_raw(value: u32) -> Self {
-        Self(value)
-    }
-
-    /// Returns the raw value of this register.
-    pub const fn raw(self) -> u32 {
-        self.0
-    }
-
-    /// Returns whether the interrupt with the given index (0-31) is pending.
-    pub const fn is_pending(self, index: u32) -> bool {
-        if index >= 32 {
-            return false;
-        }
-        (self.0 & (1 << index)) != 0
-    }
-
-    /// Sets the pending status for the interrupt with the given index (0-31).
-    pub const fn set_pending(mut self, index: u32, pending: bool) -> Self {
-        if index >= 32 {
-            return self;
-        }
-        if pending {
-            self.0 |= 1 << index;
-        } else {
-            self.0 &= !(1 << index);
-        }
-        self
-    }
-}
-
-/// External interrupt-enable bits register (`eie[n]`).
-///
-/// Each bit represents whether the corresponding interrupt identity is enabled.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct Eie(u32);
-
-impl Eie {
-    /// Creates a new `Eie` from a raw value.
-    pub const fn from_raw(value: u32) -> Self {
-        Self(value)
-    }
-
-    /// Returns the raw value of this register.
-    pub const fn raw(self) -> u32 {
-        self.0
-    }
-
-    /// Returns whether the interrupt with the given index (0-31) is enabled.
-    pub const fn is_enabled(self, index: u32) -> bool {
-        if index >= 32 {
-            return false;
-        }
-        (self.0 & (1 << index)) != 0
-    }
-
-    /// Sets the enable status for the interrupt with the given index (0-31).
-    pub const fn set_enabled(mut self, index: u32, enabled: bool) -> Self {
-        if index >= 32 {
-            return self;
-        }
-        if enabled {
-            self.0 |= 1 << index;
-        } else {
-            self.0 &= !(1 << index);
-        }
-        self
-    }
-}
-
 /// Top external interrupt register (`topei`).
 ///
 /// This register returns the interrupt identity and priority of the highest-priority
@@ -286,7 +123,8 @@ pub mod msi {
 
 /// Interrupt file operation utilities.
 pub mod file_ops {
-    use super::{Eie, Eip, MAX_INTERRUPT_IDENTITY, select};
+    use super::{MAX_INTERRUPT_IDENTITY, select};
+    use crate::csrind::{eie::Eie, eip::Eip};
 
     /// Returns `(register_index, bit_position)` for the given interrupt identity.
     #[inline]
